Removed warning error of Clang compiler build

+++ arch/arm64/include/arch/arch_ops.h
@@ -47,7 +47,7 @@ static inline void arch_disable_ints(void)
 
 static inline bool arch_ints_disabled(void)
 {
-    unsigned int state;
+    uint64_t state;
 
     __asm__ volatile("mrs %0, daif" : "=r"(state));
     state &= (1<<7);
@@ -70,7 +70,7 @@ static inline void arch_disable_fiqs(void)
 // XXX
 static inline bool arch_fiqs_disabled(void)
 {
-    unsigned int state;
+    uint64_t state;
 
     __asm__ volatile("mrs %0, daif" : "=r"(state));
     state &= (1<<6);
